add DotNetLightning.ClnRpc
This commit adds a new project for calling c-lightning from .NET
in a typesafe way.
Most of the codes are auto-generated from client generator built in
ref: https://github.com/ElementsProject/lightning/pull/5201

It must be updated after https://github.com/ElementsProject/lightning/pull/5216 gets merged.
So that all the generation logics is defined in this repo.
But this is good enough for temporary solution.